Update for C++11 standard library
This commit represents the collective efforts of the BDE team
to significantly advance the C++11 conformance of the 'bsl'
library, initially on the shared 'cpp11/baseline' branch, and
now squashed as a single (revertible) commit for the master
branch.

+++ groups/bsl/bsltf/bsltf_stdtestallocator.t.cpp
@@ -537,14 +537,14 @@ int main(int argc, char *argv[])
 
         if (verbose) printf("\tTesting 'size_type'.\n");
         {
-            ASSERT((bsl::is_same<AI::size_type, unsigned int>::value));
-            ASSERT((bsl::is_same<AV::size_type, unsigned int>::value));
+            ASSERT((bsl::is_same<AI::size_type, size_t>::value));
+            ASSERT((bsl::is_same<AV::size_type, size_t>::value));
         }
 
         if (verbose) printf("\tTesting 'difference_type'.\n");
         {
-            ASSERT((bsl::is_same<AI::difference_type, int>::value));
-            ASSERT((bsl::is_same<AV::difference_type, int>::value));
+            ASSERT((bsl::is_same<AI::difference_type, ptrdiff_t>::value));
+            ASSERT((bsl::is_same<AV::difference_type, ptrdiff_t>::value));
         }
 
         if (verbose) printf("\tTesting 'pointer'.\n");
